requirements.txt
accelerate
pillow
qwen_vl_utils
scikit-learn
spacy
torch
torchvision
tqdm
transformers
pandas
plotly
einops
tensorflow
kaleido==0.2.1
sentencepiece
protobuf
google-genai
openai
bs4
unidecode
cairosvg
IPython
datasets
# accelerate==0.24.1
# ai2-olmo==0.3.0
# aiohttp==3.9.1
# aiosignal==1.3.1
# annotated-types==0.6.0
# antlr4-python3-runtime==4.8
# anyio==4.1.0
# argon2-cffi==23.1.0
# argon2-cffi-bindings==21.2.0
# arrow==1.3.0
# asttokens==2.4.1
# async-lru==2.0.4
# async-timeout==4.0.3
# attrs==23.1.0
# Babel==2.13.1
# backoff==2.2.1
# beautifulsoup4==4.12.2
# bitsandbytes==0.42.0
# black==24.2.0
# bleach==6.1.0
# blessed==1.20.0
# blis==0.7.11
# Brotli==1.0.9
# catalogue==2.0.10
# certifi==2023.11.17
# cffi==1.16.0
# chardet==5.2.0
# charset-normalizer==3.3.2
# click==8.1.7
# cloudpathlib==0.16.0
# cmake==3.27.7
# colorama==0.4.6
# comm==0.2.0
# confection==0.1.4
# curated-tokenizers==0.0.8
# curated-transformers==0.1.1
# cycler==0.12.1
# cymem==2.0.8
# dataclasses-json==0.6.4
# dataclasses-json-speakeasy==0.5.11
# datasets==1.18.3
# debugpy==1.8.0
# decorator==5.1.1
# defusedxml==0.7.1
# dill==0.3.7
# distro==1.8.0
# einops==0.4.0
# emoji==2.10.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl
# exceptiongroup==1.2.0
# executing==2.0.1
# fairscale==0.4.13
# faiss==1.7.4
# faiss-cpu==1.7.4
# fastjsonschema==2.19.0
# filelock==3.13.1
# filetype==1.2.0
# fonttools==4.45.1
# fqdn==1.5.1
# frozenlist==1.4.0
# fsspec==2023.10.0
# gdown==4.7.1
# gpustat==1.1
# greenlet==3.0.3
# h11==0.14.0
# higher==0.2.1
# httpcore==1.0.2
# httpx==0.25.2
# huggingface-hub==0.19.4
# hydra-core==1.1.1
# idna==3.6
# importlib-metadata==6.3.0
# inquirerpy==0.3.4
# iopath==0.1.10
# ipykernel==6.27.1
# ipython==8.18.1
# ipywidgets==8.1.1
# isoduration==20.11.0
# jedi==0.19.1
# Jinja2==3.1.2
# joblib==1.3.2
# json5==0.9.14
# jsonpatch==1.33
# jsonpath-python==1.0.6
# jsonpointer==2.4
# jsonschema==4.20.0
# jsonschema-specifications==2023.11.1
# jupyter_client==8.6.0
# jupyter_core==5.5.0
# jupyter-events==0.9.0
# jupyter-lsp==2.2.1
# jupyter_server==2.11.1
# jupyter_server_terminals==0.4.4
# jupyterlab==4.0.9
# jupyterlab_pygments==0.3.0
# jupyterlab_server==2.25.2
# jupyterlab-widgets==3.0.9
# kiwisolver==1.4.5
# langchain==0.1.5
# langchain-community==0.0.17
# langchain-core==0.1.18
# langcodes==3.3.0
# langdetect==1.0.9
# langsmith==0.0.86
# lit==17.0.6
# lxml==5.1.0
# MarkupSafe==2.1.3
# marshmallow==3.20.2
# matplotlib==3.5.1
# matplotlib-inline==0.1.6
# mistune==3.0.2
# mkl-fft==1.3.1
# mkl-random==1.2.2
# mkl-service==2.4.0
# mpmath==1.3.0
# multidict==6.0.4
# multiprocess==0.70.15
# murmurhash==1.0.10
# mypy-extensions==1.0.0
# nbclient==0.9.0
# nbconvert==7.11.0
# nbformat==5.9.2
# nest-asyncio==1.5.8
# networkx==3.2.1
# nltk==3.6.5
# notebook_shim==0.2.3
# numpy==1.22.4
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cuda-cupti-cu11==11.7.101
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
# nvidia-cufft-cu11==10.9.0.58
# nvidia-curand-cu11==10.2.10.91
# nvidia-cusolver-cu11==11.4.0.1
# nvidia-cusparse-cu11==11.7.4.91
# nvidia-ml-py==12.535.133
# nvidia-nccl-cu11==2.14.3
# nvidia-nvtx-cu11==11.7.91
# omegaconf==2.1.1
# openai==1.5.0
# opencv-python==4.8.0.76
# overrides==7.4.0
# packaging==23.2
# pandas==1.4.0
# pandocfilters==1.5.0
# parso==0.8.3
# pathspec==0.12.1
# peft==0.5.0
# pexpect==4.9.0
# pfzy==0.3.4
# Pillow==10.1.0
# pip==23.3.1
# platformdirs==4.0.0
# plotly==5.18.0
# portalocker==2.8.2
# preshed==3.0.9
# prometheus-client==0.19.0
# prompt-toolkit==3.0.41
# protobuf==4.25.1
# psutil==5.9.6
# ptyprocess==0.7.0
# pure-eval==0.2.2
# pyarrow==14.0.1
# pycparser==2.21
# pydantic==2.5.2
# pydantic_core==2.14.5
# Pygments==2.17.2
# pyparsing==3.1.1
# PySocks==1.7.1
# python-dateutil==2.8.2
# python-iso639==2024.1.2
# python-json-logger==2.0.7
# python-magic==0.4.27
# pytz==2023.3.post1
# PyYAML==6.0
# pyzmq==25.1.1
# rapidfuzz==3.6.1
# referencing==0.31.1
# regex==2023.10.3
# requests==2.31.0
# rfc3339-validator==0.1.4
# rfc3986-validator==0.1.1
# rpds-py==0.13.2
# safetensors==0.4.1
# scikit-learn==1.0.2
# scipy==1.7.3
# Send2Trash==1.8.2
# sentence-transformers==2.2.2
# sentencepiece==0.1.99
# setuptools==68.0.0
# six==1.16.0
# smart-open==6.4.0
# sniffio==1.3.0
# soupsieve==2.5
# spacy==3.7.2
# spacy-alignments==0.9.1
# spacy-curated-transformers==0.2.1
# spacy-legacy==3.0.12
# spacy-loggers==1.0.5
# spacy-transformers==1.3.4
# SQLAlchemy==2.0.25
# srsly==2.4.8
# stack-data==0.6.3
# sympy==1.12
# tabulate==0.9.0
# tenacity==8.2.3
# terminado==0.18.0
# text-generation==0.6.1
# thinc==8.2.2
# threadpoolctl==3.2.0
# timm==0.9.7
# tinycss2==1.2.1
# tokenizers==0.15.0
# tomli==2.0.1
# torch==2.0.1
# torchvision==0.15.2
# tornado==6.4
# tqdm==4.62.3
# traitlets==5.14.0
# transformers==4.36.2
# triton==2.0.0
# typer==0.9.0
# types-python-dateutil==2.8.19.14
# typing_extensions==4.8.0
# typing-inspect==0.9.0
# unicode==2.9
# Unidecode==1.3.7
# unstructured==0.12.3
# unstructured-client==0.17.0
# uri-template==1.3.0
# urllib3==2.1.0
# wasabi==1.1.2
# wcwidth==0.2.12
# weasel==0.3.4
# webcolors==1.13
# webencodings==0.5.1
# websocket-client==1.6.4
# wheel==0.41.2
# widgetsnbextension==4.0.9
# wrapt==1.16.0
# xxhash==3.4.1
# yarl==1.9.3
# zipp==3.17.0
